+++ src/3d.h
@@ -15,13 +15,13 @@ typedef struct {
     bool active;
     SDL_Texture *texture;
     int state;
-    f64 y, z;
-    f64 yrot;
-    f64 xrot;
-    f64 acc, vel, jerk;
+    f32 y, z;
+    f32 yrot;
+    f32 xrot;
+    f32 acc, vel, jerk;
 
     int t, hold;
-    f64 t2;
+    f32 t2;
     int timer;
 } Object3D;
